backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
python-multipart
openpyxl
requests
python-dotenv
aiosmtplib
email-validator
